requirements.txt
grad-cam==1.4.2
numpy==1.23.1
Pillow==9.2.0
torch==1.12.1
torchvision==0.13.1
